extract Car,Distance.kt
