+++ src/main/kotlin/step3/Car.kt
@@ -1,0 +1,18 @@
+package step3
+
+data class Car(val distance: Distance = Distance()) {
+
+    init {
+        move()
+    }
+
+    var forward = 0
+
+    fun move() {
+        val currentDrive = distance.calculate()
+        when {
+            currentDrive > 4 -> forward += 1
+            else -> forward
+        }
+    }
+}
