+++ src/main/kotlin/step3/Cars.kt
@@ -18,24 +18,3 @@ data class Cars(private val numberOfCount: Int = 0) {
         }
     }
 }
-
-data class Car(val distance: Distance = Distance()) {
-
-    init {
-        move()
-    }
-
-    var forward = 0
-
-    fun move() {
-        val currentDrive = distance.calculate()
-        when {
-            currentDrive > 4 -> forward += 1
-            else -> forward
-        }
-    }
-}
-
-data class Distance(val default: Int = 0) {
-    fun calculate() = (0..9).random()
-}
