+++ src/main/kotlin/step3/Distance.kt
@@ -1,0 +1,5 @@
+package step3
+
+data class Distance(val default: Int = 0) {
+    fun calculate() = (0..9).random()
+}
